Return result object from asset creation and loading
- Create result object. This object is similar to Errorable but it also stores warnings
- Update all functions to return the Result object
- Remove all usages of assertions from the code
- Add functions that automatically load reference assets if they are not already loaded
